src/requirements.txt
Faker==2.0.0
uvicorn==0.17.6
fastapi==0.78.0
pytest==7.1.2
pydantic==1.9.1
tinydb==4.7.0
requests==2.28.0
fastapi-pagination==0.9.3
kafka-python==2.0.2
jsons==1.6.3